Add support for dataset mixtures during training (#95)
This PR makes two key changes:
1) The "data" object now contains a repeated field called "datasets"
which contains parameters specific to each dataset. The top level "data"
object still contains several parameters that should be applied across
all datasets.

2) Each dataset object exposes two new parameters for sampling and
mixing:
- `sample_count`: The number of examples to sample from the dataset.
Supports oversampling. If not specified, the entire dataset will be used
by default. Note that I did *not* add support for a sample_proportion
(percentage of the dataset) as this cannot easily be computed for
streaming datasets.
- `mixture_proportion`: The proportion of examples from this dataset
relative to other datasets in the mixture. For example, if Dataset A has
mixture_proportion=.1 and Dataset B has mixture_proportion = .9, the new
dataset will be constructed by sampling from Dataset A 10% of the time
and Dataset B 90% of the time.
 
Note that the `data` object also exposes `mixture_strategy`. This field
is the same as Hugging Face's `stopping_strategy` and dictates how
datasets are sampled when `mixture_proportion` is also defined.


Towards OPE-43

Apologies for the large PR. Changing the dataset structure caused a
breaking change in the repo.