LM Harness optimizations (#176)
Minor changes, but likely to impact performance.

1) When passing an argument as None, LM harness decides the optimal
value for this.
With this in mind:
- if `num_shots` are not set (=None), allow LM harness to decide the
optimal value.
- if `batch_size` is set to `0`, then we pass `None` to LM harness (to
decide the optimal value). Alternatively, we could make `batch_size`
optional, but this would either break other use cases or we would have
to add checks in many places (to make sure it has a meaningful value).

2) Setting `log_samples=False` so that we don't force LM harness to
return all prompts as part of the results. Though this is useful for
debugging, it will increase copying data that we don't really need most
of the time.