Support GPT2 training with Lema trainer (#302)
Towards OPE-222

Tested that GPT2 training works on 1 GPU in GCP with the changes in this
PR: https://wandb.ai/lema-academic/lema-train-test/runs/vc6g0v6h

- Updated sky config to upgrade pydantic version to prevent error when
running on GCP
- Support `cosine_with_min_lr` learning rate scheduler
- Fixed a bug where we were accidentally specifying `shuffle` for
StatefulDataLoader for iterable datasets when it should have been None.
Root cause is IterableDataset is an instance of Dataset, so we never get
to the else block intended for it.
- Due to the above change, it seems that StatefulDataLoader's state_dict
can contain tensors (`'sampler_iter_state': {'generator': tensor([1, 0,
0, ..., 0, 0, 0], dtype=torch.uint8), 'yielded': 0}},`), which isn't
JSON-serializable. I switched to using torch.save instead of serializing
to JSON, and updated the corresponding test.