[DRAFT] Clean up mixed precision params (#311)
Towards OPE-216

Ideally I would've wanted to remove the `model.torch_dtype` param so we
have one param to control both the model dtype and mixed precision
strategy, but we may still need the former on its own for
inference/eval. I'm not certain if my current approach is that much
better than what we had before, so putting out this PR for feedback.