Add a 'done' field to the LeMa job status object. (#317)
Adding a `done` field in addition to `status`. `done` is a boolean that
simply indicates if the job has terminated. It is up to each cloud to
set this value properly. `status` is still a string that can give
cloud-specific finegrain status.

The addition of this field lets us easily poll jobs in tutorials.