Minor update to Llama70B (#365)
Towards OPE-150

To get Llama70B working, you need to set `device_map="cpu"` in
`transformers.AutoModelForCausalLM.from_pretrained` in models.py.
Finding a way to programatically do this for FSDP runs will take more
experimentation, and I plan to do that in a separate PR.

We get ~60s/it with this setup, while torchtune gets 12s/it, using as
similar of a config as possible on the same cluster. It's possible
they're not packing their data, which could be the difference.