Increase from 2 to 3 nodes for Llama 70B Lora (#402)
Towards OPE-309

This is to avoid OOM issues which show up occasionally with 2 nodes on
GCP and Polaris. It's flaky with the same config, and I'm unsure of the
reason, but there's more memory breathing room with 3 nodes. However,
the downside is scheduling this job in Polaris can get much harder since
queue times are longer outside the debug queues.

This also adds in k_proj, which was in the torchtune reference config
but I didn't add in earlier PRs:
https://github.com/pytorch/torchtune/blob/main/recipes/configs/llama3_1/70B_lora.yaml