Remove flash attention validation check (#551)
Towards OPE-233, OPE-363

- Warn on unused_kwargs in HF transformers config. We don't have any
unused kwargs today, but IMO it would be helpful to catch potential
errors if we add more args in the future. This is from my experience
with the `attn_implementation` param in #533.
- Remove flash attn dtype check, as the `flash-attn` repo already does
it. If folks would still prefer we keep this, I think it should be
changed to a hard error instead of a warning.