Inference Engine async writes (#574)
This PR is based on https://github.com/oumi-ai/oumi/pull/558

This PR updates how writes are done for inference.

- If all requests are successful, the final written file will have all
responses in the same order as the provided input for all engines.
- During inference, all requests are written to a `/scratch` directory
containing a file with the same name. There is no guarantee of order on
the values written to this file (this truly only matters for the
InferenceEngines that leverage parallelism).

Non-parallel engines will write to disk in-line (blocking). I've
benchmarked this for medium sized files (100s of MB): appending a line
of text to these files takes on average 1.788e-05 seconds