Coerce model params to correct dtype for QLoRA FSDP (#652)
Towards OPE-577

This is a temporary workaround for the QLoRA FSDP bug, until there's a
fix in the transformers codebase. I had to remove the
`log_training_config` function to prevent a cyclic dependency.

Tested on Llama 3/8/70B QLora FSDP. The issue is gone on them all, and
training runs successfully on all but 3B, which has a separate issue.