Allow simpler method of specifying LoRA Adapter for eval/inference (#889)
# Description:

This PR lets us specify the adapter dir in the `model_name` field for
simplicity. Before, we had to specify the following:

`{model_name: "meta-llama/Llama-3.2-3B-Instruct", adapter_model:
"path/to/adapter"}`

Having to specify the base model isn't convenient, and could lead to
potential bugs where the wrong base model is given. With this PR, we can
now do:

`{model_name: "path/to/adapter"}`

If the base model's weights are in the same directory as the adapter,
the above is postprocessed to:

`{model_name: "path/to/adapter", adapter_model: "path/to/adapter"}`

Otherwise, the base model's name should be found in
`adapter_config.json`, and we postprocess it to:

`{model_name: "meta-llama/Llama-3.2-3B-Instruct", adapter_model:
"path/to/adapter"}`

## Implementation:

This is implemented as a postprocessing step after Config/Param
initialization. This seemed simpler and more extendable than modifying
every location that uses the `adapter_model` field; for example, the
VLLM inference engine has its own logic for loading the LoRA adapter,
and when we add a new inference engine, it's possible we'll forget to
call this method there.

The challenge with making this a postprocessing step is we can't put
this logic in ModelParam's __post_init__(). With the current
implementation of `BaseConfig.from_yaml_and_arg_list()`, the ModelParams
class is instantiated without any initial values, so the __post_init__()
function is invoked before we have the needed values.

I decided to modify the validate() function to both finalize and
validate the configs. An alternative approach is to try to modify the
logic of `from_yaml_and_arg_list()` so that __post_init__() works
properly.

## Related issues

Towards OPE-711

## Before submitting
- [ ] This PR only changes documentation. (You can ignore the following
checks in that case)
- [x] Did you read the [contributor
guideline](https://github.com/oumi-ai/oumi/blob/main/CONTRIBUTING.md)
Pull Request guidelines?
- [x] Did you link the issue(s) related to this PR in the section above?
- [x] Did you add / update tests where needed?


## Reviewers

At least one review from a member of `oumi-ai/oumi-staff` is required.

<!-- Add `oumi-ai/oumi-staff` as a reviewer when your PR is ready for
review.

You are also welcome to add individual members of `oumi-ai/oumi-staff`
as reviewers.

If no one has reviewed your PR after several days, feel free to add a
comment tagging specific reviewers.

 -->